accepted

+++ src/app/page/add-batch/add-batch.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute,Router } from '@angular/router';
 import { BatchService } from 'src/app/batch.service';
 import { CookieService } from 'ngx-cookie-service';
-
+import Swal from 'sweetalert2'; 
 
 @Component({
   selector: 'app-add-batch',
@@ -41,8 +41,17 @@ export class AddBatchComponent implements OnInit {
       });
   }
   logout() {
-    // Clear cookies and navigate to the login page
-    this.cookieService.delete('Department');
-    this.router.navigate(['/login']); // Replace '/login' with the path to your login page
+    Swal.fire({
+      position: 'center',
+      icon: 'success',
+      title: 'Logged out successfully',
+      showConfirmButton: false,
+      timer: 1500
+    }).then(() => {
+      // Clear cookies and navigate to the login page
+      this.cookieService.delete('Department');
+      this.router.navigate(['/login']); // Replace '/login' with the path to your login page
+    });
   }
+  
 }
